refactor: ranking 관련 constants 파일 분리 #11
