+++ src/constants/ranking/ageGroups.ts
@@ -1,0 +1,8 @@
+export const AGE_GROUPS = [
+  { label: '전체', value: 1 },
+  { label: '20대', value: 2 },
+  { label: '30대', value: 3 },
+  { label: '40대', value: 4 },
+  { label: '50대', value: 5 },
+  { label: '60대 이상', value: 6 },
+] as const;
